Separate "prepare cf" and "cf help"
Dedicated, ordered command classes & specs
