+++ lib/bosh/cli/commands/99_cf_help.rb
@@ -1,0 +1,16 @@
+module Bosh::Cli::Command
+  class CloudFoundryHelp < Base
+
+    usage "cf"
+    desc  "show micro cf sub-commands"
+    def cf_help
+      say("bosh cf sub-commands:")
+      nl
+      cmds = Bosh::Cli::Config.commands.values.find_all {|c|
+        c.usage =~ /cf/
+      }
+      Bosh::Cli::Command::Help.list_commands(cmds)
+    end
+
+  end
+end
